Part 01 Step 01
At first it starts from the definition of the base test with no implementations, goes on with the implementation of the structures needed to make the test compile and finishes with a dummy implementation to make the test pass.

Then the duplication introduced by the dummy implementation is removed without breaking the test.