Part 01 Step 03
Since Dollar create always a new Dollar is a *Value Object*. This kind of objects need the equals method and, if are related with a database, a hashCode method too.

In this step first was tested the equality, then the disequality to make a triangularization that helps to find the code duplication.