Part 01 Step 11
Since Dollar and Franc are not used anymore they can be deleted.

The TestDifferentClassEquality is a test for the old logic, now the subclasses are no longer used so the test is a obstacle and must be deleted.

Since there is only money some tests become redundant and can be deleted.